Relax check_version_info to check for bytecode compatibility
As discussed in #3988, the intention of check_version_info is to ensure
that cloudpickle can transfer objects across the two Python
interpreters. Because cloudpickle serializes Python bytecode, it is
known to be incompatible across minor versions (e.g., 3.11 to 3.12) but
should be compatible across patch versions (e.g., 3.12.0 to 3.12.1).

The intention of the CPython team is that bytecode should be compatible
within patch releases to the same minor version. This was last broken in
3.5.3, which was regarded as a mistake: see the commit message of
python/cpython@93602e3af70d3b9f98ae2da654b16b3382b68d50 (in 3.5.10),
which implies that it is reasonable to "rel[y] on pre-cached bytecode
remaining valid across maintenance releases."

The constant importlib.util.MAGIC_NUMBER is used by Python to identify
bytecode compatibility (it is stored in .pyc files and checked when they
are loaded). The unwanted change in 3.5.3 bumped MAGIC_NUMBER, but since
then, it has only changed between minor versions. See the long comment
in CPython's Lib/importlib/_bootstrap_external.py for a history of the
changes.

So, the practical effect of checking MAGIC_NUMBER will generally be to
enforce that nodes run the same minor version of Python, but if some
future patch release unexpectedly does break bytecode compatibility,
checking MAGIC_NUMBER will account for that.

Signed-off-by: Geoffrey Thomas <geofft@twosigma.com>